vuln_db/requirements.txt
psycopg2-binary
python-dotenv
requests
# Multi-source vulnerability system dependencies
beautifulsoup4>=4.9.0
feedparser>=6.0.0
python-dateutil>=2.8.0
lxml>=4.6.0